ms lengths as double, as they are source of truth for sample lengths

+++ include/metadata.h
@@ -9,9 +9,9 @@ struct TrackMetadata {
   TrackMetadata();
   ~TrackMetadata();
 
-  int64_t Length; // milliseconds
+  double Length; // milliseconds
   int64_t LengthSamples;
-  int64_t Fadeout; // milliseconds
+  double Fadeout; // milliseconds
   int64_t FadeoutSamples;
   std::string Title;
   std::string Artist;
